requirements.txt
python-telegram-bot==13.4.1
requests==2.25.1
PyYAML==5.4.1